Added mount adapters for Daylight/Disco on a Stick
Created some small spacers that adapt the mount hole spacing to accommodate Daylight (or Disco) on a Stick LED PCBs. As I found the 10 degree tilt for the original mounts not to my own liking, I added an additional 35 degrees of tilt in those spacers, bringing the total tilt of the LEDs to 45 degrees below horizontal. Not currently using the original diffuser bars, but I may model some kind of diffuser in the future.